Start reorganizing
* Put the 'Generically' and 'Generically1' types in a module,
  `Prelude.Linear.Generically`, at the very base of the module
  hierarchy. This seems to help avoid the module cycle problems that
  have been worked around various ways.

* Move the generic-deriving machinery for each class into the module
  that defines that class.

* Start unwinding orphan instance spaghetti. I'm optimistic this
  will work, but I know it's not a sure thing.
